some bug fixes
I added a variable and an if statement to see if the player is loading a save to stop the timer from counting it against them. I also put the quest complete splitter in an if statement to stop it from auto splitting whenever the player reloads a save or dies.